# public/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Transaction API",
        "contact": {
            "name": "DigitaleFactuur",
            "url": "https://digitalefactuur.nl/"
        },
        "version": "1.0"
    },
    "host": "transaction.test",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/states": {
            "get": {
                "tags": [
                    "state"
                ],
                "summary": "Array of states",
                "description": "Returns states array.",
                "operationId": "api.states.index",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "States[]."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/state/{id}": {
            "get": {
                "tags": [
                    "state"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns state object.",
                "operationId": "api.states.show",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of state to return",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "State overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "state"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns state overview.",
                "operationId": "api.state.delete",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "State id to delete",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "State overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "patch": {
                "tags": [
                    "state"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns state object that has been updated.",
                "operationId": "api.states.update",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of state to update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Updated user object",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string",
                                    "maximum": 64
                                },
                                "description": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "State overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/state": {
            "post": {
                "tags": [
                    "state"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns states overview.",
                "operationId": "api.state.store",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string",
                                    "maximum": 64
                                },
                                "description": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "States overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
    },
...
}